services/dispatcher/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-cloud-pubsub==2.19.0
google-cloud-storage==2.14.0
vertexai
pydantic==2.6.1